pyproject.toml
[tool.poetry]
name = "promptcraft"
version = "0.1.0"
description = ""
authors = ["shubhadeep <shubhadeeproychowdhury@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0"
click = "^8.1.3"
langchain = "^0.0.173"
openai = "^0.27.6"
semver = "^3.0.0"
rich = "^13.3.5"

[tool.poetry.scripts]
promptcraft = "promptcraft.main:main"


[tool.poetry.group.dev.dependencies]
ipython = "^8.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
